tests: add some test cases
